GUA-209: Refactor JWT creation to separate function
